+++ lib/msal-common/src/index.ts
@@ -4,6 +4,7 @@ export { PublicClientSPAConfiguration, AuthOptions } from "./app/config/PublicCl
 export { SystemOptions, LoggerOptions, TelemetryOptions } from "./app/config/ModuleConfiguration";
 // Account
 export { Account } from "./auth/Account";
+export { IdTokenClaims } from "./auth/IdTokenClaims";
 // Authority
 export { Authority } from "./auth/authority/Authority";
 export { AuthorityFactory } from "./auth/authority/AuthorityFactory";
@@ -26,6 +27,7 @@ export { CodeResponse } from "./response/CodeResponse";
 export { ILoggerCallback, LogLevel } from "./logger/Logger";
 // Errors
 export { AuthError, AuthErrorMessage } from "./error/AuthError";
+export { ServerError } from "./error/ServerError";
 export { ClientAuthError, ClientAuthErrorMessage } from "./error/ClientAuthError";
 export { ClientConfigurationError, ClientConfigurationErrorMessage } from "./error/ClientConfigurationError";
 // Constants and Utils